Split MobileSyncPage into container and component, with JSX
